+++ frontend/src/components/Templates/cave/cavev1/floor.png
